+++ tiled/adapters/csv.py
@@ -1,4 +1,5 @@
 import copy
+import inspect
 from pathlib import Path
 from typing import Any, Callable, Dict, Iterable, Iterator, List, Optional, Tuple, Union
 from urllib.parse import quote_plus
@@ -15,7 +16,15 @@ from ..structures.table import TableStructure
 from ..type_aliases import JSON
 from ..utils import ensure_uri, path_from_uri
 from .array import ArrayAdapter
-from .utils import init_adapter_from_catalog
+from .utils import filter_kwargs, init_adapter_from_catalog
+
+# Keyword arguments allowed in pandas.read_csv
+ALLOWED_KWARGS = set(
+    name
+    for name, p in inspect.signature(pandas.read_csv).parameters.items()
+    if p.kind
+    in (inspect.Parameter.KEYWORD_ONLY, inspect.Parameter.POSITIONAL_OR_KEYWORD)
+)
 
 
 class CSVAdapter:
@@ -42,10 +51,11 @@ class CSVAdapter:
         specs :
         kwargs : dict
             any keyword arguments that can be passed to the pandas.read_csv function, e.g. names, sep, dtype, etc.
+            The allowed parameters are listed in `ALLOWED_KWARGS`, which is used to filter user input.
         """
         self._file_paths = [path_from_uri(uri) for uri in data_uris]
         self._metadata = metadata or {}
-        self._read_csv_kwargs = kwargs
+        self._read_csv_kwargs = filter_kwargs(ALLOWED_KWARGS, **kwargs)
         if structure is None:
             table = dask.dataframe.read_csv(
                 self._file_paths[0], **self._read_csv_kwargs
@@ -315,9 +325,10 @@ class CSVArrayAdapter(ArrayAdapter):
         structure = data_source.structure
         dtype_numpy = structure.data_type.to_numpy_dtype()
         nrows = kwargs.pop("nrows", None)  # dask doesn't accept nrows
-        _kwargs = {"dtype": dtype_numpy, "header": None}
-        _kwargs.update(kwargs)
-        ddf = dask.dataframe.read_csv(file_paths, **_kwargs)
+        kwargs = filter_kwargs(
+            ALLOWED_KWARGS, **{"dtype": dtype_numpy, "header": None, **kwargs}
+        )
+        ddf = dask.dataframe.read_csv(file_paths, **kwargs)
         chunks_0: tuple[int, ...] = structure.chunks[
             0
         ]  # chunking along the rows dimension (when not stackable)
@@ -361,9 +372,8 @@ class CSVArrayAdapter(ArrayAdapter):
         **kwargs: Optional[Any],
     ) -> "CSVArrayAdapter":
         file_paths = [path_from_uri(uri) for uri in data_uris]
-        array = dask.dataframe.read_csv(
-            file_paths, header=None, **kwargs
-        ).to_dask_array()
+        kwargs = filter_kwargs(ALLOWED_KWARGS, **{"header": None, **kwargs})
+        array = dask.dataframe.read_csv(file_paths, **kwargs).to_dask_array()
         structure = ArrayStructure.from_array(array)
 
         return cls(array, structure)